Pull most API endpoints out into separate modules
