+++ server/dbActions.js
@@ -1,0 +1,38 @@
+/*global Promise */
+
+module.exports = function(db) {
+    var users = db.collection("users");
+    var conversations = db.collection("conversations");
+    var messages = db.collection("messages");
+    var notifications = db.collection("notifications");
+    var groups = db.collection("groups");
+
+    return {
+        cleanIdField: cleanIdField,
+        clearNotifications: clearNotifications
+    };
+
+    function clearNotifications(userID, type, data) {
+        var notificationQuery = {
+            userID: userID,
+            type: type
+        };
+        for (var key in data) {
+            notificationQuery["data." + key] = data[key];
+        }
+        notifications.deleteMany(notificationQuery);
+    }
+
+    // Returns a copy of the input with the "_id" field renamed to "id"
+    function cleanIdField(obj) {
+        var cleanObj = {};
+        for (var key in obj) {
+            if (key === "_id") {
+                cleanObj.id = obj._id;
+            } else {
+                cleanObj[key] = obj[key];
+            }
+        }
+        return cleanObj;
+    }
+};
